pyproject.toml
[tool.poetry]
name = "contentman"
version = "0.1.0"
description = ""
authors = ["user.name <gmail@hdknr.com>"]
readme = ["README.md"]

[tool.poetry.scripts]
contentman = "contentman.runner:main"


[tool.poetry.dependencies]
python = "^3.9"
django-environ = "^0.9.0"
click = "^8.1.3"
contentful-management = "^2.11.0"
contentful = "^1.13.1"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.1.0"
black = "^22.10.0"
isort = "^5.10.1"
pyproject-flake8 = "^6.0.0"
flake8-black = "^0.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
exclude = '''
(
    migrations
    | .mypy_cache
    | .pytest_cache
    | .tox
    | venv
)
'''

[tool.isort]
include_trailing_comma = true
line_length = 119
multi_line_output = 3

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,W503,W504"
exclude = "migrations"

[tool.pytest.ini_options]
django_find_project = false
addopts = "--cov --cov-report html"
filterwarnings = []

[tool.coverage.run]
omit =  []

[tool.coverage.report]
skip_covered = true
skip_empty = true
